Added Smartmontools Version 7.5 builds
Builds for CORE 13.3 and SCALE 25.04 if you want to upgrade now to the official smartmontools 7.5
Read the included readme file for instruction on how to install.
The Multi-Report User Guide version 3.17 Appendix C describes in detail how to compile your own.  It is easy with these instructions and sort of enjoyable.